Rebasing and addressing conflicts.
